==> 01.download-pubmed.ipynb <==
# Download XML efetch records for PubMed IDs in the specified journals

Uses functions from `pubmedpy` rather than the `eutils` Python package, which [does not support](https://github.com/biocommons/eutils/issues/124/) pagination.

import pathlib
import lzma
import tqdm

from pubmedpy.eutilities import esearch_query, download_pubmed_ids

%%time
# for journal in journals:
payload = {
    'db': 'pubmed',
    'term': '"journal article"[pt] AND 1993:2019[dp] AND "computational biology"[MeSH Terms] AND English[Language]',
}
pubmed_ids = esearch_query(payload, tqdm=tqdm.tqdm_notebook)
pubmed_ids = sorted(map(int, pubmed_ids))
print(f'{len(pubmed_ids):,} articles for English computational biology')

%%time
path = pathlib.Path('data/pubmed/esummary/compbio-english.xml.xz')
path.parent.mkdir(parents=True, exist_ok=True)
with lzma.open(path, 'wt') as write_file:
    download_pubmed_ids(
        pubmed_ids, write_file, endpoint='esummary',
        retmax=200, retmin=50, sleep=0, error_sleep=1,
        tqdm=tqdm.tqdm_notebook,
    )

%%time
path = pathlib.Path('data/pubmed/efetch/compbio-english.xml.xz')
path.parent.mkdir(parents=True, exist_ok=True)
with lzma.open(path, 'wt') as write_file:
    download_pubmed_ids(
        pubmed_ids, write_file, endpoint='efetch',
        retmax=200, retmin=50, sleep=0, error_sleep=1,
        tqdm=tqdm.tqdm_notebook,
    )
