==> 04.process-pmc.ipynb <==
# Extract authors from PMC-OAI frontmatter `<article>` records

import pathlib

import pandas

from pubmedpy.xml import yield_etrees_from_zip
from pubmedpy.pmc_oai import extract_authors_from_article

zip_paths = sorted(pathlib.Path('data/pmc/oai/pmc_fm').glob('*.zip'))
zip_paths

authors = list()
for zip_path in zip_paths:
    for name, article in yield_etrees_from_zip(zip_path):
        authors.extend(extract_authors_from_article(article))
author_df = pandas.DataFrame(authors)
author_df = author_df.sort_values(['pmcid', 'position'])
affiliation_df = author_df[["pmcid", "position", "affiliations"]]
author_df = author_df.drop(columns=['affiliations'])
author_df.tail()

# create affiliations table
affiliation_df = (
    affiliation_df
    .explode('affiliations')
    .rename(columns={"affiliations": "affiliation"})
    [["pmcid", "position", "affiliation"]]
    .dropna(subset=["affiliation"])
)
affiliation_df.head(2)

# Show 10 random affiliations
print(*affiliation_df.sample(10, random_state=0).affiliation, sep='\n')

# number of unique affiliations
affiliation_df.affiliation.nunique()

# Total number of articles
author_df.pmcid.nunique()

# number of corresponding authors per paper
n_corresponding = author_df.groupby("pmcid").corresponding.sum()
pmcids_without_corresponding = set(n_corresponding[n_corresponding == 0].index)

# Probability of author position being corresponding,
# given that there's at least one corresponding author
# and the author is not the last author
(
    author_df
    .query("pmcid not in @pmcids_without_corresponding")
    .query("reverse_position > 1")
    .groupby("position")
    .corresponding
    .mean()
    .map("{:.1%}".format)
    .head()
)

# Probability of author reverse position being corresponding,
# given that there's at least one corresponding author
# and the author is not the first author
(
    author_df
    .query("pmcid not in @pmcids_without_corresponding")
    .query("position > 1")
    .groupby("reverse_position")
    .corresponding
    .mean()
    .map("{:.1%}".format)
    .head()
)

# Corresponding author counts
n_corresponding.value_counts().sort_index()

# Testing: show some articles without any corresponding authors
n_corresponding.reset_index().query("corresponding == 0").head()

# Testing: show some articles without >10 corresponding authors
n_corresponding.reset_index().query("corresponding >= 10")

# Write author dataframe to a TSV
author_df.to_csv('data/pmc/authors.tsv.xz', index=False, sep='\t')

# Write affiliation dataframe to a TSV
affiliation_df.to_csv('data/pmc/affiliations.tsv.xz', index=False, sep='\t')
