==> 05.download-iscb.ipynb <==
# Download ISCB honorees (keynotes / fellows)

From [this google sheet](https://docs.google.com/spreadsheets/d/1QPjvv-8Mhk6lI720PHYsNud-MjjZbCnVQqY30f1-8hI/edit#gid=0).


import pandas

url = "https://docs.google.com/spreadsheets/d/1QPjvv-8Mhk6lI720PHYsNud-MjjZbCnVQqY30f1-8hI/export?format=csv"
keynote_df = pandas.read_csv(url)
keynote_df.head()

# Rows with any missing values, i.e. dataframe.keepna
keynote_df[keynote_df.isna().any(axis='columns')]

path = "data/iscb/keynotes.tsv"
keynote_df.to_csv(path, index=False, sep='\t')
