==> 07.corresponding-authors.ipynb <==
# Created a merged corresponding authors dataset

Identifies corresponding authors from PubMed Central author list if available. Otherwise, use PubMed author list.

import pandas

pubmed_df = pandas.read_csv("data/pubmed/articles.tsv.xz", sep="\t")
pubmed_df.head(2)

pubmed_author_df = pandas.read_csv("data/pubmed/authors.tsv.xz", sep="\t")
pubmed_author_df.head(2)

pmc_author_df = pandas.read_csv("data/pmc/authors.tsv.xz", sep="\t")
pmc_author_df.head(2)

def get_corresponding(df):
    if df.corresponding.any():
        df = df.query("corresponding == 1")
        df['use_last'] = 0
    else:
        df = df.query("reverse_position == 1")
        df['use_last'] = 1
    return df.assign(n_corresponding=len(df))

pmc_corresp_df = (
    pubmed_df[['pmid', 'pmcid']]
    .merge(pmc_author_df)
    .assign(source="pmc")
    .groupby('pmid')
    .apply(get_corresponding)
)
pmc_corresp_df.head(2)

pubmed_corresp_df = (
    pubmed_author_df
    .query("pmid not in @pmc_corresp_df.pmid")
    .assign(corresponding=None, source="pubmed")
    .groupby('pmid')
    .apply(get_corresponding)
)
pubmed_corresp_df.head(2)

corresp_df = pandas.concat([pmc_corresp_df, pubmed_corresp_df], sort=False).reset_index(drop=True)
fore_df = pandas.read_csv("data/names/fore-names.tsv.xz", sep="\t")
last_df = pandas.read_csv("data/names/last-names.tsv.xz", sep="\t")
corresp_df = (
    corresp_df
    .merge(fore_df[["fore_name", "fore_name_simple"]], how="left")
    .merge(last_df[["last_name", "last_name_simple"]], how="left")
)
corresp_df.head(2)

source_df = (
    pubmed_df
    .merge(corresp_df[['pmid', 'source', 'use_last']].drop_duplicates())
)
source_df.head(2)

# use_last prevalence
pandas.crosstab(source_df.source, source_df.use_last, margins=True)

# number of articles by authorship source by journal
pandas.crosstab(source_df.journal, source_df.source, margins=True)

# number of articles by authorship source by year
pandas.crosstab(source_df.publication_date.str.slice(0, 4), source_df.source, margins=True)

corresp_df.to_csv('data/names/corresponding-authors.tsv.xz', sep='\t', index=False)
