==> 08.corresponding-authors.ipynb <==
# Created a merged corresponding authors dataset

Identifies corresponding authors from PubMed Central author list if available. Otherwise, use PubMed author list.

import pathlib
import pandas
import pubmedpy

## Prepare country assignment

# read affiliations to country mapping
country_df = pandas.read_csv("data/affiliations/countries.tsv.xz", sep='\t', keep_default_na=False)
# be careful reading country codes: Namibia code is NA
assert country_df.notna().all(axis=None)
country_df.head()

source = 'pubmed'
directory = pathlib.Path('data') / source
affil_df = pandas.read_csv(directory / "affiliations.tsv.xz", sep="\t")
affil_df = affil_df.merge(country_df)


def read_authors(source: str) -> pandas.DataFrame:
    """
    Read authors table for source, and add a countries column.
    """
    directory = pathlib.Path('data') / source
    # read authors
    author_df = pandas.read_csv(directory / "authors.tsv.xz", sep="\t")
    # read author affiliations
    affil_df = pandas.read_csv(directory / "affiliations.tsv.xz", sep="\t")
    affil_df = affil_df.merge(country_df)
    # assign countries to authors by affiliations
    primary_key = {"pubmed": "pmid", "pmc": "pmcid"}[source]
    countries_df = (
        affil_df
        .groupby([primary_key, "position"])
        .country
        .apply(lambda x: ",".join(sorted(set(x))))
        .reset_index()
        .rename(columns={"country": "countries"})
    )
    author_df = author_df.merge(countries_df, how='left')
    return author_df

## Read pubmed authors

pubmed_df = pandas.read_csv("data/pubmed/articles.tsv.xz", sep="\t")
pubmed_df.head(2)

pubmed_df.head()

pubmed_author_df = read_authors("pubmed")
pubmed_author_df.head(2)

# pmc_author_df = read_authors("pmc")
# pmc_author_df.head(2)

# def get_corresponding(df):
# #     if df.corresponding.any():
# #         df = df.query("corresponding == 1")
# #         df['use_last'] = 0
# #     else:
#     df = df.query("reverse_position == 1")
#     df['use_last'] = 1
#     return df.assign(n_corresponding=len(df))

# pmc_corresp_df = (
#     pubmed_df[['pmid', 'pmcid']]
#     .merge(pmc_author_df)
#     .assign(source="pmc")
#     .groupby('pmid')
#     .apply(get_corresponding)
# )
# pmc_corresp_df.head(2)

# pmc_corresp_df.countries.notna().mean()

pubmed_author_df.query('reverse_position == 1')

# pubmed_corresp_df.countries.notna().mean()

# corresp_df = pandas.concat([pmc_corresp_df, pubmed_corresp_df], sort=False).reset_index(drop=True)
corresp_df = (pubmed_author_df
    .query('reverse_position == 1')
    .assign(source='pubmed', use_last = 1))
fore_df = pandas.read_csv("data/names/fore-names.tsv.xz", sep="\t", keep_default_na=False)
last_df = pandas.read_csv("data/names/last-names.tsv.xz", sep="\t", keep_default_na=False)
corresp_df = (
    corresp_df
    .merge(fore_df[["fore_name", "fore_name_simple"]], how="left")
    .merge(last_df[["last_name", "last_name_simple"]], how="left")
)
corresp_df.head(2)

# number of authors with 1+ assinged countries by source
# False indicates no assigned countries.
# True indicates one or more countries
tab = pandas.crosstab(
    corresp_df.source,
    corresp_df.countries.notnull(),
    margins=True,
)

# percent of authors with 1+ assinged countries by source
pandas.crosstab(
    corresp_df.source,
    corresp_df.countries.notnull(),
    margins=True, normalize="index"
).applymap("{:.1%}".format)

pmc_country = (
    pubmed_df
    .merge(corresp_df[['pmid', 'source', 'countries']].drop_duplicates())
)
pmc_country.head()

If available, collect PMCID of articles with no countries to query more affiliations from:

pmcids_query = (pmc_country.query('(countries != countries) and (pmcid == pmcid)'))['pmcid']

pmcids_query

def get_frontmatter_etree_via_api(pmcid):
    url = f"https://www.ncbi.nlm.nih.gov/pmc/oai/oai.cgi?verb=GetRecord&identifier=oai:pubmedcentral.nih.gov:{pmcid[3:]}&metadataPrefix=pmc_fm"
    response = requests.get(url)
    tree = etree.fromstring(response.content)
    article = tree.find("{*}GetRecord/{*}record/{*}metadata/{*}article")
    return article

from pubmedpy.pmc_oai import extract_authors_from_article
# from pubmedpy.tests.test_pmc_oai import get_frontmatter_etree_via_api
from lxml import etree
# from lxml.etree import tostring
import requests

art = get_frontmatter_etree_via_api('PMC6986237')
extract_authors_from_article(art)

art = get_frontmatter_etree_via_api('PMC6986235')
print(etree.tostring(art, encoding = 'unicode'))

extract_authors_from_article(art)

### Analyze sources for corresponding authors

source_df = (
    pubmed_df
    .merge(corresp_df[['pmid', 'source', 'use_last']].drop_duplicates())
)
source_df.head(2)

# number of articles by authorship source by journal
pandas.crosstab(source_df.journal, source_df.source, margins=True)

# number of articles by authorship source by year
pandas.crosstab(source_df.publication_date.str.slice(0, 4), source_df.source, margins=True)

# Number of articles where all corresponding authors have assigned countries by journal
articles_with_na_countries = set(corresp_df.loc[corresp_df.countries.isna()].pmid)
source_df["corresp_has_countries"] = ~source_df.pmid.isin(articles_with_na_countries)
pandas.crosstab(source_df.journal, source_df.corresp_has_countries, margins=True)

# Percent of articles where all corresponding authors have assigned countries by journal
pandas.crosstab(
    source_df.journal, source_df.corresp_has_countries,
    margins=True, normalize="index"
).applymap("{:.1%}".format)

corresp_df.to_csv('data/names/corresponding-authors.tsv.xz', sep='\t', index=False)
