==> 08.visualize-gender.Rmd <==
---
title: "Visualize gender predictions based on census data"
output: html_document
---

## Setups

```{r message = F, warning=F}
library(tidyverse)
library(lubridate)
source('utils/r-utils.R')
theme_set(theme_bw() + theme(legend.title = element_blank()))
```

## Load data

```{r}
alpha_threshold <- qnorm(0.975)

simple_fore_names <- read_tsv('data/names/fore-names.tsv.xz') %>% 
  select(- n_authors) %>% 
  distinct()
gender_df <- read_tsv('data/gender/genderize.tsv')

pubmed_gender_pmids <- read_tsv('data/names/corresponding-authors.tsv.xz') %>%
  left_join(read_tsv('data/pubmed/articles.tsv.xz'), by = 'pmid') %>%
  mutate(year = substr(publication_date, 1, 4) %>% ymd(truncated = 2),
         publication_date = ymd(publication_date, truncated = 2)) %>%
  left_join(gender_df, by = 'fore_name_simple')

pubmed_gender_df <- pubmed_gender_pmids %>% 
  group_by(pmid, journal, publication_date, year) %>% 
  summarise(probability_male = mean(probability_male, na.rm = T)) %>% 
  ungroup()

pubmed_gender_pmids %>% filter(year < '2002-01-01', !is.na(fore_name_simple))

iscb_gender_df <- read_tsv('data/iscb/keynotes.tsv') %>%
  mutate(publication_date = ymd(year, truncated = 2),
         year = ymd(year, truncated = 2)) %>% 
  left_join(simple_fore_names, by = 'fore_name') %>% 
  left_join(gender_df, by = 'fore_name_simple') %>% 
  mutate(final_gender_pred = ifelse(probability_male > 0.5, 'Male', 'Female'))

start_year <- 1997
end_year <- 2020
n_years <- end_year - start_year
my_jours <- unique(pubmed_gender_df$journal)
my_confs <- unique(iscb_gender_df$conference)
n_jours <- length(my_jours)
n_confs <- length(my_confs)

# 294 unique names in iscb dataset:
iscb_gender_df %>% 
  distinct(full_name) %>% 
  nrow() 

```

34050 *PMC plus fall back* corresponding author names.
29749 articles.

Between 2000-2001, 11 corresponding author names from 10 BMC Bioinformatics articles were retrieved from PMC.


Number of speakers/fellows in compbio conferences across years:

```{r}
iscb_gender_df %>%
  select(year, conference) %>% 
  count(year, conference) %>% 
  ggplot(aes(x = year(year), y = n, color = conference)) +
  geom_point() +
  geom_line(alpha = 0.5) +
  coord_cartesian(ylim = c(0, 13)) +
  scale_x_continuous(breaks = seq(1998, 2020, 2)) +
  scale_y_continuous(breaks = seq(0, 14, 2)) +
  scale_color_viridis_d() +
  labs(x = 'Year', y = 'Number of keynote speakers/fellows')
```

## Missing data

```{r}
iscb_gender_df %>% filter(is.na(probability_male))
pubmed_gender_pmids %>% filter(is.na(probability_male)) %>% sample_n(10)
sum(is.na(pubmed_gender_df$probability_male))
```

2 iscb forenames not genderized (Árpád and Chung-I).
2835 pubmed forenames not genderized.
These mostly include initials only, hyphenated names and perhaps names with accent marks.

```{r}
pubmed_gender_df %>% filter(is.na(probability_male)) %>% 
  count(year, journal) %>% 
  ggplot(aes(year(year), n, fill = journal)) +
  geom_bar(stat = 'identity') +
  scale_fill_viridis_d() +
  ylab('Number of forenames NOT genderized')

pubmed_gender_pmids %>%
  filter(is.na(probability_male),
         journal == 'BMC Bioinformatics',
         year < '2002-01-01')
```

The metadata for Bioinformatics and BMC Bioinformatics papers before 2002 only have initials of first and/or middle names. 
Therefore, we do not have prediction for genders before this year.

## Descriptive statistics
Prepare data frames for later analyses:

- rbind results of race predictions in iscb and Pubmed
<!-- - pivot long -->
- compute mean, sd, marginal error

```{r}
iscb_pubmed <- iscb_gender_df %>%
  rename('journal' = conference) %>% 
  select(year, journal, probability_male, publication_date) %>%
  mutate(type = 'Keynote speakers/Fellows') %>%
  bind_rows(
    pubmed_gender_df %>%
      select(year, journal, probability_male, publication_date) %>%
      mutate(type = 'Pubmed authors')
  ) %>%
  mutate(probability_female = 1 - probability_male) %>% 
  pivot_longer(contains('probability'),
               names_to = 'gender',
               values_to = 'probabilities') %>%
  group_by(type, year, publication_date, gender) %>%
  add_count() %>%
  mutate(
    mean_prob = mean(probabilities, na.rm = T),
    sd_prob = sd(probabilities, na.rm = T),
    n = mean(n),
    me_prob = alpha_threshold * sd_prob / sqrt(n)
  ) %>%
  ungroup() 
```

## Prepare data frames for analysis
### By journals

```{r}
pubmed_gender <- vector('list', length = n_jours)
i <- 0

for (jour in my_jours){
  i <- i + 1
  pubmed_gender[[i]] <- iscb_pubmed %>%
    filter(type == 'Pubmed authors' & journal == jour) %>%
    group_by(year, gender, journal) %>%
    summarise(mean_prob = mean(probabilities, na.rm = T)) %>%
    ungroup() 
}



```

### By conference keynotes/fellows
```{r fig.height=6}
i <- 0
iscb_gender <- vector('list', length = n_confs)

for (conf in my_confs){
  i <- i + 1
  iscb_gender[[i]] <- iscb_pubmed %>%
    filter(type != 'Pubmed authors' & journal == conf) %>%
    group_by(year, gender, journal) %>%
    summarise(mean_prob = mean(probabilities, na.rm = T)) %>%
    ungroup()
}
```

## Comparisons

```{r}
fig_1a <- iscb_pubmed %>%
  filter(type == 'Pubmed authors') %>%
  group_by(year, type, gender) %>%
  summarise(mean_prob = mean(probabilities, na.rm = T)) %>%
  ungroup() %>%
  gender_breakdown('main', type)

fig_1b <- iscb_pubmed %>%
  filter(type != 'Pubmed authors') %>%
  group_by(year, type, gender) %>%
  summarise(mean_prob = mean(probabilities, na.rm = T)) %>%
  ungroup() %>%
  gender_breakdown('main', type)

fig_1c <- bind_rows(iscb_gender) %>%
  gender_breakdown(category = 'sub', journal) +
  theme(legend.position = c(0.2, 0.95),
        legend.key.height = unit(4, 'mm'),
        legend.background = element_rect(fill = "transparent", colour = "transparent"))

left_col <- cowplot::plot_grid(fig_1a, fig_1b, labels = c('A', 'B'), nrow = 2)
fig_1 <- cowplot::plot_grid(left_col, fig_1c, labels = c('', 'C'), ncol = 2)
fig_1
ggsave('figs/gender_breakdown.png', fig_1, width = 7, height = 5)

# By journals:
bind_rows(pubmed_gender) %>%
  gender_breakdown(category = 'sub', journal)

```

## Overview of gender prediction distribution

```{r fig.height = 3}
iscb_pubmed %>% 
  filter(gender == 'probability_male') %>% 
  ggplot(aes(x = probabilities, y = ..density.., fill = type)) +
  geom_histogram(position = 'dodge') +
  scale_fill_viridis_d(option = 'D', end = 0.8) +
  guides(label = F) +
  labs(x = 'probability_male')
```

## Mean and standard deviation of predicted probabilities

A simpler figure, supplementary Fig S1?

```{r}
ggplot(iscb_pubmed %>% filter(gender == 'probability_male'),
  aes(group = type)) +
  geom_smooth(size = 0.2, data = . %>% filter(type == 'Pubmed authors'),
              aes(x = publication_date, y = probabilities, fill = type, color = type)) +
  geom_pointrange(data = . %>% get_keynote_summary(),
                  alpha = 0.75, size = 0.3,
                  aes(x = year, y = mean_prob, shape = type,
                      ymin = mean_prob - me_prob,
                      ymax = mean_prob + me_prob)) +
  scale_y_continuous(breaks = seq(0, 1, 0.2), labels = scales::percent_format()) +
  scale_x_date(labels = scales::date_format("%Y")) +
  scale_color_viridis_d(option = 'D', begin = 0.5) +
  scale_fill_viridis_d(option = 'D', begin = 0.5) +
  coord_cartesian(ylim = c(0, 1)) +
  labs(x = NULL, y = 'Estimated composition') +
  theme(legend.position = c(0.2, 0.2),
        legend.margin = margin(-0.5, 0, 0, 0, unit='cm'))
```

```{r}
sessionInfo()
```
